refactor(model): add utils/class_utils.py
