+++ wenet/branchformer/cgmlp.py
@@ -25,7 +25,7 @@ References:
 from typing import Tuple
 import torch
 import torch.nn as nn
-from wenet.utils.common import get_activation
+from wenet.utils.class_utils import get_activation
 
 
 class ConvolutionalSpatialGatingUnit(torch.nn.Module):
