+++ wenet/squeezeformer/encoder.py
@@ -29,7 +29,7 @@ from wenet.squeezeformer.positionwise_feed_forward \
     import PositionwiseFeedForward
 from wenet.squeezeformer.convolution import ConvolutionModule
 from wenet.utils.mask import make_pad_mask, add_optional_chunk_mask
-from wenet.utils.common import get_activation
+from wenet.utils.class_utils import get_activation
 
 
 class SqueezeformerEncoder(nn.Module):
